Move CasperEngine to its own package
