+++ casper/src/main/scala/coop/rchain/casper/engine/BlockApproverProtocol.scala
@@ -1,4 +1,4 @@
-package coop.rchain.casper.util.comm
+package coop.rchain.casper.engine
 
 import cats.Traverse
 import cats.data.EitherT
